# backend/core/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma de cinema screeK.",
        "title": "screeK API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte screeK",
            "url": "http://www.screek.com/support",
            "email": "suporte@screek.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8003",
    "basePath": "/",
    "paths": {
...
        "/bookings/admin/sessions/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancela todos os ingressos de uma sessão em caso de problemas técnicos no cinema",
                "tags": [
                    "Admin"
                ],
                "summary": "Cancelar sessão por contingência (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sessão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/bookings/admin/sessions/{id}/tickets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna todos os ingressos vendidos para uma sessão específica",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Listar ingressos da sessão (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da Sessão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bookings.TicketResponseDTO"
                            }
                        }
                    }
                }
            }
        },
        "/bookings/admin/tickets/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancela um ingresso sem restrição de tempo ou dono",
                "tags": [
                    "Admin"
                ],
                "summary": "Cancelar ingresso (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Ticket (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/bookings/playing": {
            "get": {
                "description": "Retorna todos os filmes que possuem sessões ativas na cidade e data informada",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Listar filmes em cartaz",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cidade do cinema",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data (format: YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bookings.MovieResponseDTO"
                            }
                        }
                    }
                }
            }
        },
        "/bookings/sessions/{id}/seats": {
            "get": {
                "description": "Retorna todos os assentos de uma sessão e seu status de ocupação\nRetorna todos os assentos de uma sessão e seu status de ocupação",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings",
                    "Bookings"
                ],
                "summary": "Mapa de assentos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da sessão",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bookings.SeatResponseDTO"
                            }
                        }
                    }
                }
            }
        },
        "/bookings/tickets/reserve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma reserva temporária de 10 minutos para os assentos selecionados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Reservar assentos",
                "parameters": [
                    {
                        "description": "Dados da reserva",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookings.ReserveRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bookings.ReserveResponseDTO"
                        }
                    }
                }
            }
        },
        "/bookings/tickets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna informações detalhadas de um ingresso específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Detalhe do ingresso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Ticket (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bookings.TicketResponseDTO"
                        }
                    }
                }
            }
        },
        "/bookings/tickets/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancela um ingresso e processa estorno se aplicável",
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancelar ingresso",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Ticket (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httputil.MessageResponse"
                        }
                    }
                }
            }
        },
        "/bookings/transactions/{id}/pay": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Processa o pagamento de uma reserva pendente via Stripe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Pagar reserva",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da Transação (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chave de Idempotência",
                        "name": "idempotency-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Método de pagamento",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bookings.PayRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bookings.PayResponseDTO"
                        }
                    }
                }
            }
        },
        "/bookings/users/me/tickets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o histórico de ingressos do usuário autenticado",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Meus ingressos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtrar por status (PAID, PENDING, CANCELLED)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bookings.TicketResponseDTO"
                            }
                        }
                    }
                }
            }
        },
        "/bookings/{id}/sessions": {
            "get": {
                "description": "Retorna as sessões de um filme agrupadas por cinema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Consultar sessões de um filme",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do filme (TMDB)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Cidade",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/bookings.CinemaSessionsResponseDTO"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bookings.CinemaSessionsResponseDTO": {
            "type": "object",
            "properties": {
                "cinema_city": {
                    "type": "string"
                },
                "cinema_id": {
                    "type": "integer"
                },
                "cinema_name": {
                    "type": "string"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bookings.SessionResponseDTO"
                    }
                }
            }
        },
        "bookings.MovieResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_premiere": {
                    "type": "boolean"
                },
                "is_rescreening": {
                    "type": "boolean"
                },
                "poster_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tmdb_id": {
                    "type": "integer"
                }
            }
        },
        "bookings.PayRequestDTO": {
            "type": "object",
            "required": [
                "payment_method"
            ],
            "properties": {
                "payment_method": {
                    "type": "string"
                }
            }
        },
        "bookings.PayResponseDTO": {
            "type": "object",
            "properties": {
                "client_secret": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "bookings.ReserveRequestDTO": {
            "type": "object",
            "required": [
                "session_id",
                "tickets_request"
            ],
            "properties": {
                "session_id": {
                    "type": "integer"
                },
                "tickets_request": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/bookings.TicketRequest"
                    }
                }
            }
        },
        "bookings.ReserveResponseDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "valor_total_centavos": {
                    "type": "integer"
                }
            }
        },
        "bookings.SeatResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_occupied": {
                    "type": "boolean"
                },
                "number": {
                    "type": "integer"
                },
                "pos_x": {
                    "type": "integer"
                },
                "pos_y": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "row": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "bookings.SessionResponseDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "room_type": {
                    "type": "string"
                },
                "session_type": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "bookings.TicketRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "seat_id": {
                    "type": "integer"
                },
                "type": {
                    "enum": [
                        "STANDARD",
                        "HALF",
                        "FREE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/bookings.TicketType"
                        }
                    ]
                }
            }
        },
        "bookings.TicketResponseDTO": {
            "type": "object",
            "properties": {
                "cinema": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "movie_name": {
                    "type": "string"
                },
                "qr_code": {
                    "type": "string"
                },
                "room": {
                    "type": "string"
                },
                "seat": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "ticket_id": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/bookings.UserBookingDTO"
                }
            }
        },
        "bookings.TicketType": {
            "type": "string",
            "enum": [
                "STANDARD",
                "HALF",
                "FREE"
            ],
            "x-enum-varnames": [
                "TicketTypeStandard",
                "TicketTypeHalf",
                "TicketTypeFree"
            ]
        },
        "bookings.UserBookingDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "httputil.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}